pyproject.toml
[project]
name = "vehicle-ai-agent"
version = "2.0.0"
description = "Klasyfikacja pojazdów (MobileNetV2) + agent NL→SQL — zadanie rekrutacyjne Atlasus"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.30",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "Pillow>=10.0.0",
    "openai>=1.40.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "streamlit>=1.38.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [".venv_before_rename_fix", ".pytest_cache_before_rename_fix"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
